session-manager/pyproject.toml
[project]
name = "session-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-openid-auth>=0.17",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "python-multipart>=0.0.20",
    "python3-openid>=3.2.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
